+++ src/fortune_wheel/fortune_wheel_command.tsx
@@ -62,7 +62,7 @@ export class FWCommandScreen extends MBSObjectScreen<FWCommand> {
 
         const disabled = !character.IsPlayer();
         document.body.appendChild(
-            <div id={ID.root} class="HideOnPopup mbs-screen" screen-generated={this.screen}>
+            <div id={ID.root} class="mbs-screen">
                 <style id={ID.styles}>{styles.toString()}</style>
 
                 <h1 id={ID.header}>{`Customize wheel of fortune command ${this.index}`}</h1>