+++ src/fortune_wheel/fortune_wheel_select.tsx
@@ -128,7 +128,7 @@ export class FWSelectScreen extends MBSScreen {
         const isPlayer = this.character.IsPlayer();
 
         document.body.appendChild(
-            <div id={ID.root} class="HideOnPopup mbs-screen" screen-generated={this.screen}>
+            <div id={ID.root} class="mbs-screen">
                 <style id={ID.styles}>{styles.toString()}</style>
 
                 <div id={ID.buttonOuterGrid}>