+++ src/garbling/gag_speak.ts
@@ -1,15 +1,7 @@
-import { range } from "lodash-es";
+import { range, zip } from "lodash-es";
 
 import { MBS_MOD_API } from "../common";
-
-import _GAG_DATA from "./vendor/gag_data.json" assert { type: "JSON" };
-import _PHONETIC_DICT from "./vendor/en_UK.json" assert { type: "JSON" };
-
-type PhoneticDict = Readonly<Record<string, undefined | string>>;
-type GagData = Readonly<Record<string, { readonly MUFFLE: number, readonly SOUND: string }>>;
-
-const GAG_DATA: { readonly [k in keyof typeof _GAG_DATA]: GagData } = Object.freeze(_GAG_DATA);
-const PHONETIC_DICT: PhoneticDict = Object.freeze(_PHONETIC_DICT);
+import { json, GagData } from "./json";
 
 export namespace GarbleOptions {
     export interface DropChars {
@@ -30,6 +22,7 @@ export namespace GarbleOptions {
         readonly character?: Character,
         readonly fallback?: Fallback,
         readonly ignoreOOC?: boolean,
+        readonly perSyllable?: boolean,
     }
 
     export interface Parsed extends Record<keyof Base, unknown> {
@@ -37,20 +30,25 @@ export namespace GarbleOptions {
         readonly character: null | Character,
         readonly fallback: Fallback,
         readonly ignoreOOC: boolean,
+        readonly perSyllable: boolean,
     }
 }
 
-export const GAG_LEVEL = Object.freeze({
-    light: 3,
-    medium: 6,
-    heavy: 9,
-    max: 20,
+type GagLevel = typeof GagLevel[keyof typeof GagLevel];
+export const GagLevel = Object.freeze({
+    LIGHT: 3,
+    MEDIUM: 6,
+    HEAVY: 9,
+    KILO_HEAVY: 12,
+    MEGA_HEAVY: 15,
+    GIGA_HEAVY: 18,
+    MAX: 21,
 });
 
-const PHONEMES_SYMBOLS_EN_UK = Object.freeze(new Set([
-    "b", "aʊ", "t", "k", "ə", "z", "ɔ", "ɹ", "s", "j", "u", "m", "f", "ɪ", "oʊ", "ʊ", "ɡ", "ɛ", "n",
-    "eɪ", "d", "ɫ", "w", "i", "p", "ɑ", "ɝ", "θ", "v", "h", "æ", "ŋ", "ʃ", "ʒ", "aɪ", "dʒ", "tʃ", "ð",
-    "ɔɪ", "ɪə",
+const PHONEMES_SYMBOLS_EN_US = Object.freeze(new Set([
+    "aɪ", "aʊ", "b", "d", "dʒ", "eɪ", "f", "h", "i", "j", "k", "m", "n", "oʊ", "p", "s", "t",
+    "tʃ", "u", "v", "w", "z", "æ", "ð", "ŋ", "ɑ", "ɔ", "ɔɪ", "ə", "ɛ", "ɝ", "ɡ", "ɪ", "ɪə", "ɫ",
+    "ɹ", "ʃ", "ʊ", "ʒ", "θ",
 ]));
 
 const LETTER_REGEX = /\p{L}/u;
@@ -58,27 +56,29 @@ const LETTER_REGEX = /\p{L}/u;
 /**
  * Convert a single English word to a list of its respective International Phonetic Alphabet (IPA) characters
  * @param word The to-be converted word
- * @returns The words IPA characters or `undefined` if the word is unknown
+ * @returns The words IPA characters per syllable or `undefined` if the word is unknown
  */
-function wordToIPA(word: string): undefined | string[] {
+function wordToIPA(word: string): undefined | string[][] {
     const wordLower = word.toLowerCase();
-    const wordPhonetic = PHONETIC_DICT[wordLower];
-    if (wordPhonetic === undefined) {
+    const phoneticSyllables = json.phoneticDict[wordLower];
+    if (phoneticSyllables === undefined) {
         return undefined;
     }
 
-    const phonetics: string[] = [];
-    const wordPhoneticSan = wordPhonetic.split(",")[0].replace(/[/ˌˈː\s]/g, "");
-    let i = 0;
-    while (i < wordPhoneticSan.length) {
-        const char1 = wordPhoneticSan[i];
-        const char12 = wordPhoneticSan.slice(i, i + 2);
-        if (PHONEMES_SYMBOLS_EN_UK.has(char12)) {
-            phonetics.push(char12);
-            i += 2;
-        } else {
-            phonetics.push(char1);
-            i += 1;
+    const phonetics: string[][] = range(0, phoneticSyllables.length).map(_ => []);
+    for (const items of zip(phoneticSyllables, phonetics)) {
+        const [syllable, phoneticsList] = items as [string, string[]];
+        let i = 0;
+        while (i < syllable.length) {
+            const char1 = syllable[i];
+            const char12 = syllable.slice(i, i + 2);
+            if (PHONEMES_SYMBOLS_EN_US.has(char12)) {
+                phoneticsList.push(char12);
+                i += 2;
+            } else {
+                phoneticsList.push(char1);
+                i += 1;
+            }
         }
     }
     return phonetics;
@@ -102,8 +102,8 @@ function parseOptions(options: null | GarbleOptions.Base, gagLevel: number): Gar
     if (options?.dropChars != null) {
         const dropCharsOptions: GarbleOptions.DropChars = {
             level: gagLevel,
-            levelMin: options.dropChars.levelMin ?? GAG_LEVEL.heavy,
-            levelMax: options.dropChars.levelMax ?? GAG_LEVEL.max,
+            levelMin: options.dropChars.levelMin ?? GagLevel.HEAVY,
+            levelMax: options.dropChars.levelMax ?? GagLevel.MAX,
             maxFrac: options.dropChars.maxFrac ?? 0.5,
         };
         dropChars = (word) => dropTrailingChars(word, dropCharsOptions);
@@ -114,19 +114,60 @@ function parseOptions(options: null | GarbleOptions.Base, gagLevel: number): Gar
         character: options?.character ?? null,
         fallback: options?.fallback ?? ((args) => MBS_MOD_API.callOriginal("SpeechGarbleByGagLevel", args)),
         ignoreOOC: options?.ignoreOOC ?? false,
+        perSyllable: options?.perSyllable ?? false,
     };
 }
 
-function getGagData(character: null | Character, gagLevel: number): GagData {
+function getGagData(
+    character: null | Character,
+    gagLevel: number,
+    syllableOptions: null | { readonly syllable: number, readonly syllableMax: number } = null,
+): GagData {
     if (character && character.HasEffect("OpenMouth") && !character.HasEffect("BlockMouth")) {
-        return GAG_DATA["Ring Gag"];
-    } else if (gagLevel <= GAG_LEVEL.light) {
-        return GAG_DATA["Cleave Gag"];
-    } else if (gagLevel <= GAG_LEVEL.medium) {
-        return GAG_DATA["Ball Gag"];
-    } else {
-        return GAG_DATA["Dildo Gag"];
+        return json.gagData["Ring Gag"] ?? {};
     }
+
+    syllableOptions: if (syllableOptions !== null) {
+        const gagLevelRound = Math.floor(gagLevel / 3);
+        if (gagLevelRound === 0) {
+            break syllableOptions;
+        }
+
+        const frac = (gagLevel / 3) - gagLevelRound;
+        if (frac > (syllableOptions.syllable / syllableOptions.syllableMax)) {
+            gagLevel = (1 + gagLevelRound) * 3;
+        }
+    }
+
+    if (gagLevel <= GagLevel.LIGHT) {
+        return json.gagData["Cleave Gag"] ?? {};
+    } else if (gagLevel <= GagLevel.MEDIUM) {
+        return json.gagData["Ball Gag"] ?? {};
+    } else {
+        return json.gagData["Dildo Gag"] ?? {};
+    }
+}
+
+/** Cut a sentence into OOC and non-OOC chunks */
+function sentenceCutOOC(sentence: string, ignoreOOC: boolean = false): { words: string, ooc: boolean }[] {
+    const oocIter = SpeechGetOOCRanges(sentence).values();
+    let oocNext = oocIter.next();
+    if (oocNext.done || ignoreOOC) {
+        return [{ words: sentence, ooc: false }];
+    }
+
+    const ret: { words: string, ooc: boolean}[] = [];
+    let start = 0;
+    while (!oocNext.done) {
+        const oocStart = oocNext.value.start;
+        const oocStop = oocStart + oocNext.value.length;
+        ret.push({ words: sentence.slice(start, oocStart), ooc: false });
+        ret.push({ words: sentence.slice(oocStart, oocStop), ooc: true });
+        oocNext = oocIter.next();
+        start = oocStop;
+    }
+    ret.push({ words: sentence.slice(start), ooc: false });
+    return ret;
 }
 
 export function convertToGagSpeak(
@@ -134,39 +175,40 @@ export function convertToGagSpeak(
     gagLevel: number,
     options: null | GarbleOptions.Base = null,
 ): string {
-    const { dropChars, fallback, character, ignoreOOC } = parseOptions(options, gagLevel);
-    const gagData = getGagData(character, gagLevel);
-
-    const oocIndices = new Set(SpeechGetOOCRanges(sentence).flatMap(({ start, length }) => {
-        return range(start, start + length + 1);
-    }));
+    const { dropChars, fallback, character, ignoreOOC, perSyllable } = parseOptions(options, gagLevel);
 
     let garbledSentence = "";
-    let word = "";
-    let caps: boolean[] = [];
-    let inOOC = false;
-    const iMax = sentence.length - 1;
-    for (const [i, char] of Array.from(sentence).entries()) {
-        if (LETTER_REGEX.test(char) && i < iMax) {
-            word += char;
-            inOOC ||= oocIndices.has(i);
-            caps.push(char.toUpperCase() === char);
+    const sentenceFragments = sentenceCutOOC(sentence, ignoreOOC);
+    for (const sentenceFrag of sentenceFragments) {
+        // Got an OOC fragment here; no need to garble anything
+        if (sentenceFrag.ooc) {
+            garbledSentence += sentenceFrag.words;
             continue;
         }
 
-        if (!word) {
-            garbledSentence += fallback([gagLevel, char]);
-            continue;
-        }
+        let word = "";
+        let caps: boolean[] = [];
+        const iMax = sentenceFrag.words.length - 1;
+        for (const [i, char] of Array.from(sentenceFrag.words).entries()) {
+            // Keep constructing  as long as unicode letter characters are coming through
+            const charIsLetter = LETTER_REGEX.test(char);
+            if (charIsLetter) {
+                word += char;
+                caps.push(char.toUpperCase() === char);
+                if (i < iMax) {
+                    continue;
+                }
+            }
 
-        if (inOOC && !ignoreOOC) {
-            // No Need for garbling while in OOC
-            garbledSentence += word;
-        } else {
             // Try the phonetics-based garbling and fall back to BC's default if no match is found for the word
-            const phonetics = wordToIPA(word);
-            if (phonetics) {
-                const garbledWord = dropChars(phonetics.map(j => gagData[j]?.SOUND ?? "").join(""));
+            const phonWord = wordToIPA(word);
+            if (phonWord) {
+                const garbledWord = dropChars(phonWord.flatMap((phonSyllable, j) => {
+                    const syllableOptions = perSyllable ? { syllable: j + 1, syllableMax: phonWord.length } : null;
+                    const gagData = getGagData(character, gagLevel, syllableOptions);
+                    return phonSyllable.map(phonChar => gagData[phonChar]?.SOUND ?? "");
+                }).join(""));
+
                 if (caps.every(Boolean)) {
                     garbledSentence += garbledWord.toUpperCase();
                 } else if (caps[0] && garbledWord.length > 0) {
@@ -175,14 +217,16 @@ export function convertToGagSpeak(
                     garbledSentence += garbledWord;
                 }
             } else {
+                garbledSentence += fallback([gagLevel, word]);
+            }
+
+            word = "";
+            caps = [];
+            if (!charIsLetter) {
+                // Avoid double garbling `char` if it's a letter
                 garbledSentence += fallback([gagLevel, char]);
             }
         }
-
-        inOOC = false;
-        word = "";
-        caps = [];
-        garbledSentence += char;
     }
     return garbledSentence;
 }