+++ src/garbling/index.ts
@@ -2,17 +2,28 @@ import { MBS_MOD_API, waitFor, logger } from "../common";
 import { bcLoaded } from "../common_bc";
 
 import { convertToGagSpeak, GarbleOptions } from "./gag_speak";
+import { json as garblingJSON } from "./json";
+
+export { garblingJSON };
 
 waitFor(bcLoaded).then(() => {
     logger.log("Initializing garbling module");
 
+    // Only use this hook for ring gags
     MBS_MOD_API.hookFunction("SpeechGarble", 0, ([character, words, noDeaf, ...args], next) => {
         const gagLevel = SpeechGetTotalGagLevel(character, noDeaf);
-        if (gagLevel === 0 || !character || !words || !Player.MBSSettings.AlternativeGarbling) {
+        if (
+            !gagLevel
+            || !character
+            || !words
+            || !Player.MBSSettings.AlternativeGarbling
+            || !(character.HasEffect("OpenMouth") && !character.HasEffect("BlockMouth"))
+        ) {
             return next([character, words, noDeaf, ...args]);
         } else {
             const options: GarbleOptions.Base = {
                 dropChars: Player.MBSSettings.DropTrailing ? {} : undefined,
+                perSyllable: Player.MBSSettings.GarblePerSyllable,
                 character,
             };
             let ret = convertToGagSpeak(words, gagLevel, options);
@@ -30,6 +41,7 @@ waitFor(bcLoaded).then(() => {
         } else {
             const options: GarbleOptions.Base = {
                 dropChars: Player.MBSSettings.DropTrailing ? {} : undefined,
+                perSyllable: Player.MBSSettings.GarblePerSyllable,
                 fallback: next,
                 ignoreOOC,
             };