+++ src/settings/settings_screen.tsx
@@ -14,6 +14,7 @@ import {
     SettingsStatus,
     clearMBSSettings,
 } from "./settings";
+import { garblingJSON } from "../garbling";
 
 import styles from "./settings_screen.scss";
 
@@ -111,7 +112,7 @@ export class MBSPreferenceScreen extends MBSScreen {
         super(parent, MBSPreferenceScreen.screenParamsDefault, params);
 
         document.body.appendChild(
-            <div id={ID.root} class="HideOnPopup mbs-screen" screen-generated={this.screen}>
+            <div id={ID.root} class="mbs-screen">
                 <style id={ID.styles}>{styles.toString()}</style>
 
                 <h1 id={ID.header}>{`Maid's Bondage Scripts ${MBS_VERSION}`}</h1>
@@ -164,13 +165,21 @@ export class MBSPreferenceScreen extends MBSScreen {
                                 <strong>Experimental</strong>: whether gags will use an alternative form of, more phonetically acurate, speech garbling
                                 based on <a href="https://github.com/CordeliaMist/Dalamud-GagSpeak" target="_blank">Dalamud-GagSpeak</a>
                             </p>
-                            <p>Incompatible-ish with FBC's garbling anti-cheat as of the moment</p>
+                            <p>
+                                Incompatible-ish with <a href="https://sidiousious.gitlab.io/bce/" target="_blank">FBC</a>'s
+                                garbling anti-cheat as of the moment
+                            </p>
                         </div>
                     </div>
                     <div class="mbs-preference-settings-pair">
                         <input type="checkbox" data-field="DropTrailing" onClick={this.#boolSwitch.bind(this)}/>
                         Whether to heaviest gags will drop up to half of all trailing characters
                         when alternate garbling is enabled
+                    </div>
+                    <div class="mbs-preference-settings-pair">
+                        <input type="checkbox" data-field="GarblePerSyllable" onClick={this.#boolSwitch.bind(this)}/>
+                        Interpolate between the three alternative garbling levels, allowing for a more gradual increase
+                        in garbling strength (on a syllable by syllable basis) as the gag level increases
                     </div>
                 </div>
 
@@ -225,7 +234,7 @@ export class MBSPreferenceScreen extends MBSScreen {
         );
 
         document.body.appendChild(
-            <div id={ID.resetScreen} class="HideOnPopup mbs-screen" screen-generated={MBSPreferenceScreen.screen}>
+            <div id={ID.resetScreen} class="mbs-screen" screen-generated={MBSPreferenceScreen.screen}>
                 <div id={ID.resetBackground}>
                     <h1>MBS data reset</h1>
                     <p><strong>- Warning -</strong></p>
@@ -266,6 +275,9 @@ export class MBSPreferenceScreen extends MBSScreen {
             Player.MBSSettings[field] = !Player.MBSSettings[field];
             pushMBSSettings([SettingsType.SETTINGS]);
             switch (field) {
+                case "AlternativeGarbling":
+                    if (Player.MBSSettings[field]) garblingJSON.init();
+                    break;
                 case "LockedWhenRestrained":
                     this.#lockInputs();
                     break;
@@ -408,13 +420,20 @@ waitFor(bcLoaded).then(() => {
         preferenceLoadHook();
     });
 
-    MBS_MOD_API.hookFunction("ServerPlayerIsInChatRoom", 0, (args, next) => {
-        if (CurrentScreen == MBSPreferenceScreen.screen && InformationSheetPreviousScreen == "ChatRoom") {
-            return true;
-        } else {
-            return next(args);
-        }
-    });
+    if (typeof ServerPlayerChatRoom !== "undefined") {
+        ServerPlayerChatRoom.register({
+            screen: MBSPreferenceScreen.screen,
+            callback: () => InformationSheetPreviousScreen === "ChatRoom",
+        });
+    } else {
+        MBS_MOD_API.hookFunction("ServerPlayerIsInChatRoom", 0, (args, next) => {
+            if (CurrentScreen == MBSPreferenceScreen.screen && InformationSheetPreviousScreen == "ChatRoom") {
+                return true;
+            } else {
+                return next(args);
+            }
+        });
+    }
 
     MBS_MOD_API.hookFunction("PreferenceClick", 0, (args, next) => {
         const previousScreen = PreferenceSubscreen;
@@ -427,7 +446,9 @@ waitFor(bcLoaded).then(() => {
 
     (PreferenceSubscreenList as string[]).push(MBSPreferenceScreen.screen);
     preferenceState = new PreferenceScreenProxy();
-    if (CurrentScreen === "Preference") {
-        preferenceLoadHook();
+
+    switch (CurrentScreen) {
+        case "Preference":
+            return preferenceLoadHook();
     }
 });